% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandicoot_tree-data.R
\docType{data}
\name{bandicoot_tree}
\alias{bandicoot_tree}
\title{Phylogenetic tree of living and recently extinct bandicoots}
\format{
An object of the class \code{"phylo"}
}
\usage{
data(bandicoot_tree)
}
\description{
A dated and fully resolved phylogenetic tree of living and recently extinct
bandicoots (Perameloidea). Reproduced from the tree published in Kear, Aplin &
Westerman (2016).
}
\references{
Kear BP, Aplin KP & Westerman M (2016) Bandicoot fossils and DNA
  elucidate lineage antiquity amongst xeric-adapted Australasian marsupials.
  Scientific Reports 6: 37537.
}
\keyword{datasets}
