% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrington_birds-data.R
\docType{data}
\name{barrington_birds}
\alias{barrington_birds}
\title{A community data table of bird counts, Barrington Tops National Park}
\format{
A \code{data.frame} with a row for each site and a column for each
  species. Values are counts of each species in each site.
}
\usage{
data(barrington_birds)
}
\description{
A table of counts of birds across sites in Barrington Tops National Park,
NSW, Australia. Data were collected in November 2012 from 29 points where all
birds observed over a 30 minute period within a 50 metre radius were
recorded.
}
\references{
Chao A, Chiu CH, Hsieh TC, Davis T, Nipperess DA & Faith DP
  (2016) Rarefaction and extrapolation of phylogenetic diversity.
  \emph{Methods in Ecology and Evolution} 6: 380-388.
}
\keyword{datasets}
