% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylocurve.R
\name{phylocurve}
\alias{phylocurve}
\title{Generate a rarefaction curve of Phylogenetic Diversity}
\usage{
phylocurve(x, phy, stepm = 1, subsampling = "individual", replace = FALSE)
}
\arguments{
\item{x}{is a community \code{data.frame} (as in the \code{vegan} package)
with species/OTUs as columns and samples/sites as rows. Columns are
labelled with the names of the species/OTUs. Rows are labelled with the
names of the samples/sites. Data can be either abundance or incidence
(0/1). Column labels must match tip labels in the phylogenetic tree
exactly!}

\item{phy}{is a rooted phylogenetic tree with branch lengths stored as a 
phylo object (as in the \code{ape} package) with terminal nodes labelled
with names matching those of the community data table. Note that the
function trims away any terminal taxa not present in the community data
table, so it is not necessary to do this beforehand.}

\item{stepm}{is the size of the interval in a sequence of numbers of
individuals, sites or species to which \code{x} is to be rarefied.}

\item{subsampling}{indicates whether the subsampling will be by
\code{'individual'} (default), \code{'site'} or \code{'species'}. When
there are multiple sites, rarefaction by individuals or species is done by
first pooling the sites.}

\item{replace}{is a boolean indicating whether subsampling should be done
with or without replacement.}
}
\value{
a matrix object of three columns giving the expected PD values (mean
  and variance) for each value of \code{m}
}
\description{
Calculates a rarefaction curve giving expected phylogenetic diversity (mean
and variance) for multiple values of sampling effort. Sampling effort can be
defined in terms of the number of individuals, sites or species. Expected
phylogenetic diversity is calculated using an exact analytical formulation
(Nipperess & Matsen 2013) that is both more accurate and more computationally
efficient than randomisation methods.
}
\references{
\itemize{
\item{Hurlbert (1971) The nonconcept of Species Diversity: a critique and
alternative parameters. \emph{Ecology} 52: 577-586.}
\item{Nipperess & Matsen (2013) The mean and variance of phylogenetic 
diversity under rarefaction. \emph{Methods in Ecology & Evolution} 4:
566-572.}}
}

