% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodiv_estimate.R
\name{phylodiv.estimate}
\alias{phylodiv.estimate}
\title{Estimate Phylogenetic Diversity from a sample}
\usage{
phylodiv.estimate(x, phy, subsampling = "individual")
}
\arguments{
\item{x}{is the community data given as a \code{data.frame} or \code{matrix}
with species/OTUs as columns and samples/sites as rows (like in the
\code{vegan} package). Columns are labelled with the names of the
species/OTUs. Rows are labelled with the names of the samples/sites. Data
can be either abundance or incidence (0/1). Column labels must match tip
labels in the phylogenetic tree exactly!}

\item{phy}{is a rooted phylogenetic tree with branch lengths stored as a
phylo object (as in the \code{ape} package) with terminal nodes labelled
with names matching those of the community data table. Note that the
function trims away any terminal taxa not present in the community data
table, so it is not necessary to do this beforehand.}

\item{subsampling}{indicates whether the subsampling will be by
\code{'individual'} (default) or \code{'site'}.}
}
\value{
a \code{list}, giving observed and estimated PD values. Each element
  of the list is a \code{numeric}, giving PD values for every sample/site or a
  single value for pooled samples/sites.
}
\description{
Estimates total phylogenetic diversity (both detected and undetected) for
each sample/site or a set of samples/sites. Estimated phylogenetic diversity
is calculated using an extrapolation of the rarefaction curve (Chao et al.
2015). For undetected phylogenetic diversity, this is considered a lower
bound.
}
\details{
\code{phylodiv_estimate} takes community data and a rooted
  phylogenetic tree (with branch lengths) and calculates estimated total
  Phylogenetic Diversity (PD) for both detected and undetected species. When
  there are multiple sites in the community data and subsampling is by site,
  data are converted to incidence (if not already) and then pooled. When
  subsampling is by individual, estimated PD is calculated for every
  sample/site.
}
\references{
\itemize{ \item{Chao et al. (2015) Rarefaction and extrapolation
  of phylogenetic diversity. \emph{Methods in Ecology & Evolution} 6:
  380-388.}}
}
