% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodiv_expect_secured.R
\name{phylodiv.expect.secured}
\alias{phylodiv.expect.secured}
\title{Expected Phylogenetic Diversity from conservation of threatened species}
\usage{
phylodiv.expect.secured(
  phy,
  species = phy$tip.label,
  managed = species,
  survival,
  feasibility = 1,
  success = 0.95
)
}
\arguments{
\item{phy}{is a rooted phylogenetic tree with branch lengths stored as a
phylo object (as in the \code{ape} package).}

\item{species}{is an optional \code{character} vector of species names (tree
will be trimmed to match). Each species must match to a tip label in
\code{phy}. By default, all tip labels in \code{phy} will be included.}

\item{managed}{is a \code{character} vector of species names to be managed
for conservation. Each name must match to a name in \code{species}. By
default, all names in \code{species} will be included.}

\item{survival}{is a \code{numeric} vector of survival probabilities, given
no management for conservation, corresponding to each species in
\code{species}.}

\item{feasibility}{is an optional \code{numeric} vector of probabilities of
successfully managing each species in \code{managed}. By default,
probability of success is assumed to be 1.}

\item{success}{is a \code{numeric} vector of length 1, giving the probability
of survival of a species if successfully managed for conservation. Value
must be less than 1.}
}
\value{
A \code{numeric} vector giving three values corresponding to:
  expected PD with management; expected PD without management; and
  the gain in expected PD with management (i.e. the difference).
}
\description{
Calculates expected Phylogenetic Diversity (PD) with and without successful
management of a list of threatened species. Note that this function uses that
version of PD that always includes the path to the root of the tree.
}
\details{
\code{phylodiv.expected.secured} takes the probability of survival
  of species (before and after management) and a phylogenetic tree (rooted
  and with branch lengths) and calculates expected Phylogenetic Diversity
  (PD). PD is defined as the total length of all branches spanning a set of
  species (Faith, 1992). Expected PD is the summed branch length expected to
  survive given the probability of survival of species (Witting & Loeschcke
  1995; Faith 2013). An optional argument (\code{feasibility}) allows for the
  gain of conserving any species to be down-weighted by the probability that
  management will be successful.
}
\examples{
data(bandicoot_tree)
threatened <- c("Macrotis_lagotis","Rhynchomeles_prattorum","Peroryctes_broadbenti",
       "Isoodon_auratus","Perameles_gunnii","Perameles_bougainville")
probs <- c(0.01,0.90,0.01,0.30,0.99,0.99,0.99,0.99,0.99,0.99,0.30,0.99,0.90,0.99,0.99,0.99,0.90,0.01,0.90)
phylodiv.expect.secured(bandicoot_tree,managed=threatened,survival=probs)

}
\references{
\itemize{\item{Faith DP. 1992. Conservation evaluation and
  phylogenetic diversity. \emph{Biological Conservation} 61: 1-10.}
  \item{Faith DP. 2013. Biodiversity and evolutionary history: useful
  extensions of the PD phylogenetic diversity assessment framework.
  \emph{Annals of the New York Academy of Sciences} 1289: 69-89.}
  \item{Witting L. & Loeschcke V. 1995. The optimization of biodiversity
  conservation. \emph{Biological Conservation} 71: 205-207.}}
}
