% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodiv_expected.R
\name{phylodiv.expected}
\alias{phylodiv.expected}
\title{Expected Phylogenetic Diversity of ecological samples}
\usage{
phylodiv.expected(x, phy, occur = TRUE)
}
\arguments{
\item{x}{is the community data given as a \code{data.frame} or \code{matrix} 
with species/OTUs as columns and samples/sites as rows (like in the 
\code{vegan} package). Columns are labelled with the names of the 
species/OTUs. Rows are labelled with the names of the samples/sites. Data 
are expressed as probabilities (ranging 0-1) of either occurrence or 
extinction. Column labels must match tip labels in the phylogenetic tree 
exactly!}

\item{phy}{is a rooted phylogenetic tree with branch lengths stored as a 
phylo object (as in the \code{ape} package) with terminal nodes labelled 
with names matching those of the community data table. Note that the 
function trims away any terminal taxa not present in the community data 
table, so it is not necessary to do this beforehand.}

\item{occur}{is a \code{logical} indicating whether \code{x} is a matrix of 
probabilities of occurrence (\code{TRUE} - default) or extinction 
(\code{FALSE})}
}
\value{
A vector of expected Phylogenetic Diversity (PD) values for each 
  sample/site in \code{x}.
}
\description{
Calculates the expected Phylogenetic Diversity (PD) of multiple samples 
simultaneously based on probabilities of occurrence (or extinction) for each 
combination of species and site. Note that this function uses that version of
PD that always includes the path to the root of the tree.
}
\details{
\code{phylodiv.expected} takes probability data and a phylogenetic 
  tree (rooted and with branch lengths) and calculates the expected 
  Phylogenetic Diversity (PD) of all samples/sites. PD is defined as the 
  total length of all branches spanning a set of terminal taxa representing 
  an ecological sample (Faith, 1992). Expected PD is the summed branch length
  expected for each site given the probability of occurrence (or extinction) 
  of species (and ancestral branches) at that site (Witting & Loeschcke 1995;
  Faith 2013). Please note that, if the common ancestor (node) of the set of 
  taxa of a sample is not the root of the tree, then the set of branches 
  connecting this node to the root are also included in the calculation.
}
\examples{
x <- matrix(data=c(0.8,0.6,0.2,0.2,0.9,0.1,0.1,0.8,0.9),nrow=3,ncol=3)
colnames(x) <- c("A","B","C")
phy <- read.tree(text="((A:1,B:1):1,C:2);")
phylodiv.expected(x,phy,occur=TRUE)
}
\references{
\itemize{\item{Faith DP. 1992. Conservation evaluation and 
  phylogenetic diversity. \emph{Biological Conservation} 61: 1-10.} 
  \item{Faith DP. 2013. Biodiversity and evolutionary history: useful 
  extensions of the PD phylogenetic diversity assessment framework.
  \emph{Annals of the New York Academy of Sciences} 1289: 69–89.} 
  \item{Witting L. & Loeschcke V. 1995. The optimization of biodiversity
  conservation. \emph{Biological Conservation} 71: 205–207.}}
}
