% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodiv.R
\name{phylodiv}
\alias{phylodiv}
\title{Phylogenetic Diversity of ecological samples}
\usage{
phylodiv(x, phy)
}
\arguments{
\item{x}{is a community \code{data.frame} (as in the \code{vegan} package)
with species/OTUs as columns and samples/sites as rows. Columns are
labelled with the names of the species/OTUs. Rows are labelled with the
names of the samples/sites. Data can be either abundance or incidence
(0/1). Column labels must match tip labels in the phylogenetic tree
exactly!}

\item{phy}{is a rooted phylogenetic tree with branch lengths stored as a 
phylo object (as in the \code{ape} package) with terminal nodes labelled
with names matching those of the community data table. Note that the
function trims away any terminal taxa not present in the community data
table, so it is not necessary to do this beforehand.}
}
\value{
A vector of Phylogenetic Diversity (PD) values for each sample/site
  in \code{x}.
}
\description{
Calculates the Phylogenetic Diversity (PD) of multiple samples
simultaneously. Note that this function uses that version of PD that always
includes the path to the root of the tree.
}

