% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloendemism.R
\name{phyloendemism}
\alias{phyloendemism}
\title{Phylogenetic Endemism of ecological samples}
\usage{
phyloendemism(x, phy, weighted = TRUE)
}
\arguments{
\item{x}{is the community data given as a \code{data.frame} or \code{matrix} 
with species/OTUs as columns and samples/sites as rows (like in the 
\code{vegan} package). Columns are labelled with the names of the 
species/OTUs. Rows are labelled with the names of the samples/sites. Data 
can be either abundance or incidence (0/1). Column labels must match tip 
labels in the phylogenetic tree exactly!}

\item{phy}{is a rooted phylogenetic tree with branch lengths stored as a 
phylo object (as in the \code{ape} package) with terminal nodes labelled 
with names matching those of the community data table. Note that the 
function trims away any terminal taxa not present in the community data 
table, so it is not necessary to do this beforehand.}

\item{weighted}{is a \code{logical} indicating whether weighted endemism 
(default) or strict endemism should be calculated.}
}
\value{
A \code{vector} object giving the phylogenetic endemism of all 
  sample/sites in \code{x}.
}
\description{
Calculates phylogenetic endemism (sum of 'unique' branch lengths) of multiple
ecological samples.
}
\details{
Takes a community data table and a rooted phylogenetic tree (with 
  branch lengths) and calculates either strict or weighted endemism in 
  Phylogenetic Diversity (PD). Strict endemism equates to the total amount of
  branch length found only in the sample/site and is described by Faith et 
  al. (2004) as PD-endemism. Weighted endemism calculates the "spatial 
  uniqueness" of each branch in the tree by taking the reciprocal of its 
  range, multiplying by branch length and summing for all branch lengths 
  present at a sample/site. Range is calculated simply as the total number of
  samples/sites at which the branch is present. This latter approach is 
  described by Rosauer et al. (2009) as \emph{Phylogenetic endemism}.
}

