% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloprunr_prob.R
\name{phyloprunr.prob}
\alias{phyloprunr.prob}
\title{Pruning algorithm to maximise gains in expected Phylogenetic Diversity from
successfully managing threatened species}
\usage{
phyloprunr.prob(
  phy,
  species = phy$tip.label,
  managed = species,
  retain = NULL,
  survival,
  feasibility = 1,
  cost = 1,
  success = 0.95,
  random = FALSE
)
}
\arguments{
\item{phy}{is a rooted phylogenetic tree with branch lengths stored as a
phylo object (as in the \code{ape} package).}

\item{species}{is an optional \code{character} vector of species names (tree
will be trimmed to match). Each species must match to a tip label in
\code{phy}. By default, all tip labels in \code{phy} will be included.}

\item{managed}{is a \code{character} vector of species names to be included
on the candidate list for conservation. Each name must match to a name in
\code{species}. By default, all names in \code{species} will be included as
candidates.}

\item{retain}{is an optional \code{character} vector of species names that
are never pruned from the candidate list (i.e. always assumed to be
conserved). Each name must match to a name in \code{managed}.}

\item{survival}{is a \code{numeric} vector of survival probabilities, given
no management for conservation, corresponding to each species in
\code{species}.}

\item{feasibility}{is an optional \code{numeric} vector of probabilities of
successfully managing each species in \code{managed}. By default,
probability of success is assumed to be 1.}

\item{cost}{is an optional \code{numeric} vector of costs of conservation
corresponding to each species in \code{managed}. Default value is 1 for all
candidate species.}

\item{success}{is a \code{numeric} vector of length 1, giving the probability
of survival of a species if successfully managed for conservation. Value
must be less than 1.}

\item{random}{is a \code{logical} vector of length 1, indicating whether
species should be pruned from the candidate list at random (\code{TRUE}),
rather than optimally (\code{FALSE}).}
}
\value{
A dataframe giving the names of each species pruned (in order of
  pruning), the expected Phylogenetic Diversity \emph{after} pruning that species,
  and the total cost of management (summed across managed species) \emph{after}
  pruning that species.
}
\description{
A pruning algorithm that seeks to maximise gains in expected Phylogenetic
Diversity (PD) from successfully managing threatened species from a candidate
list. Starting with all candidate species being assumed to be successfully
managed, the algorithm proceeds by stepwise removal of species from the
candidate list providing the smallest gain (from conservation) at that
step. The pruned candidate list at each step will be optimal for conserving
Phylogenetic Diversity for that number of species. The reverse order in which
species are pruned gives the priority for conservation.
}
\details{
\code{phyloprunr.prob} takes a phylogenetic tree (rooted and with
  branch lengths) and a list of candidate species for conservation and
  determines the optimal set of species that will maximise expected
  Phylogenetic Diversity (PD) for that set size, if those species were
  successfully managed for conservation. The algorithm determines step-wise
  which candidate species provides the smallest gain to expected PD and then
  removes it from the candidate list, adjusting the probabilities of survival
  of each branch in the tree as it goes. The algorithm will continue until
  the list of candidates is exhausted. At each step, ties are resolved by
  choosing at random from the available options. For this reason, the
  algorithm may need to be repeated to generate several equally optimal sets.
  An optional argument (\code{feasibility}) allows for the gain of conserving
  any species to be down-weighted by the probability that management will be
  successful. A further optional argument (\code{cost}) allows for gains to
  be divided by their respective costs.
}
\examples{
data(bandicoot_tree)
threatened <- c("Macrotis_lagotis","Rhynchomeles_prattorum","Peroryctes_broadbenti",
       "Isoodon_auratus","Perameles_gunnii","Perameles_bougainville")
probs <- c(0.01,0.90,0.01,0.30,0.99,0.99,0.99,0.99,0.99,0.99,0.30,0.99,0.90,0.99,0.99,0.99,0.90,0.01,0.90)
phyloprunr.prob(bandicoot_tree,managed=threatened,survival=probs)
}
\references{
\itemize{ \item{Minh B., Klaere S. & Haeseler A. (2006).
  Phylogenetic Diversity within Seconds. \emph{Systematic Biology} 55:
  769-773.}}
}
