% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylorebelo.R
\name{phylorebelo}
\alias{phylorebelo}
\title{Heuristic algorithm optimising the accumulation of Phylogenetic Diversity
from aggregating sites}
\usage{
phylorebelo(x, phy, iterations = 1)
}
\arguments{
\item{x}{is the community data given as a \code{data.frame} or \code{matrix} 
with species/OTUs as columns and samples/sites as rows (like in the 
\code{vegan} package). Columns are labelled with the names of the 
species/OTUs. Rows are labelled with the names of the samples/sites. Data 
can be either abundance or incidence (0/1). Column labels must match tip 
labels in the phylogenetic tree exactly!}

\item{phy}{is a rooted phylogenetic tree with branch lengths stored as a 
phylo object (as in the \code{ape} package) with terminal nodes labelled 
with names matching those of the community data table. Note that the 
function trims away any terminal taxa not present in the community data 
table, so it is not necessary to do this beforehand.}

\item{iterations}{is the number of iterations of the algorithm (to resolve 
ties). Default is 1.}
}
\value{
A list of two numeric matrices. The first matrix gives the optimised 
  sequence (rank) of the samples as rows, with a column for each iteration of
  the algorithm. The second matrix is gives the corresponding gains in PD for
  each sample (row) and iteration.
}
\description{
A heuristic algorithm that seeks to prioritise sites for conservation by 
optimising the accumulation of Phylogenetic Diversity (PD). PD will acumulate
most rapidly when the samples are pooled in a particular sequence such that 
the gain in PD of each additional sample is maximised.
}
\details{
\code{phylorebelo} takes community data and a phylogenetic tree 
  (rooted and with branch lengths) and calculates the optimal sequence of 
  samples for accumulating Phylogenetic Diversity (PD). The algorithm starts 
  by first selecting the sample with the highest PD. It will then choose the 
  sample most complementary to the first (giving the largest gain in PD). 
  Additional samples are chosen based on their complementarity to the set 
  already chosen until all sites have been selected. At each step, ties are 
  resolved by choosing at random from the available equally complementary 
  options. The algorithm is an adaptation of that described by Rebelo & 
  Siegfried (1992). The phylogenetic version implemented here is that used by
  Asmyhr et al. (2014).
}
\references{
\itemize{ \item{Asmyhr MG, Linke S, Hose G, & Nipperess DA. 2014.
  Systematic Conservation Planning for Groundwater Ecosystems Using 
  Phylogenetic Diversity. \emph{PLoS One} 9: e115132} \item{Rebelo AG & 
  Siegfried WR. 1992. Where should nature reserves be located in the Cape 
  Floristic Region, South Africa? Models for the spatial configuration of a 
  reserve network aimed at maximizing the protection of floral diversity. 
  \emph{Conservation Biology} 6: 243–252.}}
}
